+++ packages/react-conformance-make-styles/src/matchers/toContainClassNameLastInCalls.ts
@@ -18,8 +18,6 @@ export const toContainClassNameLastInCalls: jest.CustomMatcher = (
   result: jest.Mock<{}, MergeClassesParams>['mock']['calls'],
   className: string,
 ) => {
-  let hasCallsWithClassName: boolean = false;
-
   let pass = true;
   let message = '';
 
@@ -27,16 +25,6 @@ export const toContainClassNameLastInCalls: jest.CustomMatcher = (
     const indexInClassList = classesList.indexOf(className);
 
     if (indexInClassList >= 0) {
-      if (hasCallsWithClassName) {
-        pass = false;
-        message = [
-          `There multiple calls of mergeClasses() that contain "${className}".`,
-          `Last call: ${classesList.join(' ')}`,
-        ].join(' ');
-
-        return;
-      }
-
       if (classesList.indexOf(className, indexInClassList + 1) !== -1) {
         pass = false;
         message = [
@@ -46,8 +34,6 @@ export const toContainClassNameLastInCalls: jest.CustomMatcher = (
 
         return;
       }
-
-      hasCallsWithClassName = true;
 
       if (indexInClassList !== classesList.length - 1) {
         pass = false;