+++ packages/react-conformance-make-styles/CHANGELOG.md
@@ -1,8 +1,22 @@
 # Change Log - @fluentui/react-conformance-make-styles
 
-This log was last generated on Wed, 01 Sep 2021 07:39:56 GMT and should not be manually modified.
+This log was last generated on Thu, 02 Sep 2021 07:36:46 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [9.0.0-alpha.5](https://github.com/microsoft/fluentui/tree/@fluentui/react-conformance-make-styles_v9.0.0-alpha.5)
+
+Thu, 02 Sep 2021 07:36:46 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-conformance-make-styles_v9.0.0-alpha.4..@fluentui/react-conformance-make-styles_v9.0.0-alpha.5)
+
+### Patches
+
+- Bump @fluentui/react-conformance to v0.4.5 ([PR #19590](https://github.com/microsoft/fluentui/pull/19590) by olfedias@microsoft.com)
+
+### Changes
+
+- make number of calls configurable ([PR #19598](https://github.com/microsoft/fluentui/pull/19598) by olfedias@microsoft.com)
+- fix mergeClasses() mock ([PR #19594](https://github.com/microsoft/fluentui/pull/19594) by olfedias@microsoft.com)
 
 ## [9.0.0-alpha.4](https://github.com/microsoft/fluentui/tree/@fluentui/react-conformance-make-styles_v9.0.0-alpha.4)