+++ packages/react-provider/CHANGELOG.md
@@ -1,8 +1,21 @@
 # Change Log - @fluentui/react-provider
 
-This log was last generated on Wed, 01 Sep 2021 07:39:56 GMT and should not be manually modified.
+This log was last generated on Thu, 02 Sep 2021 07:36:46 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [9.0.0-alpha.72](https://github.com/microsoft/fluentui/tree/@fluentui/react-provider_v9.0.0-alpha.72)
+
+Thu, 02 Sep 2021 07:36:46 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-provider_v9.0.0-alpha.71..@fluentui/react-provider_v9.0.0-alpha.72)
+
+### Patches
+
+- Bump @fluentui/react-conformance to v0.4.5 ([PR #19590](https://github.com/microsoft/fluentui/pull/19590) by olfedias@microsoft.com)
+
+### Changes
+
+- rework className handling ([PR #19589](https://github.com/microsoft/fluentui/pull/19589) by olfedias@microsoft.com)
 
 ## [9.0.0-alpha.71](https://github.com/microsoft/fluentui/tree/@fluentui/react-provider_v9.0.0-alpha.71)